This should work

+++ src/Hooks/PlayLayer.hpp
@@ -47,7 +47,7 @@ class $modify(ModPlayLayer, PlayLayer) {
 		unsigned int m_activeSaveLayer = 0;
 		unsigned int m_saveLayerCount = 0;
 
-		std::optional<size_t> m_levelStringHash;
+		std::optional<uint64_t> m_levelStringHash;
 	};
 
 	// Hooks
